sed: Move 'cycle?' flag into the environment.
* gash/commands/sed.scm (<env>): Add 'cycle?' field.
(execute-function): When aborting to the 'end-of-script-tag', only
pass the updated environment and update the 'cycle?' field if
necessary.
(execute-commands): When aborting to the 'end-of-script-tag', only
pass the updated environment.
(edit-stream): Initialize the 'cycle?' field; only accept the
environment in the prompt handler; only loop when the 'cycle?' field
is set.