awk: Redesign the AST.
This is a rather large change.  It replaces the old, implicit AST with
one that is explicitly specified in a new file.  To build confidence
that the parser generates an AST that adheres to the specification,
many tests are added.  As a result of adding the tests, a handful of
parser bugs or missing features were exposed, all of which are
addressed by this change.

* doc/awk-syntax.txt: New file.
* Makefile.am (EXTRA_DIST): Add it.
* gash/commands/awk/lexer.scm (*operators*): Add '%='.
* gash/commands/awk/parser.scm (strip-progn): New procedure.
(unwrap-singleton): New procedure.
(make-parser): Generate an AST according to the new specification;
generate usable function declaration nodes; fix pattern ranges;
handle multi-dimensional array keys everywhere.
* tests/unit/awk-parser.scm: Add tests.
* gash/commands/awk.scm (awk-name): Delete procedure.
(awk-set, awk-expression, eval-special-items, %eval-awk): Consume
the new AST.
(run-commands): Do not handle item-level nodes; consume the new AST;
do not skip unknown commands.
(load-file, read-record, make-default-env): Use symbols for variable
names.
(eval-item): New procedure.
(run-awk-file): Use it instead of 'run-commands'.