ls: Use 'parse-options'.
This fixes a test failure on Guile 3.0.  The test removed by this
change is not the one that was failing.  The removed test is a very
low value one that simply runs 'ls --version'.  Now that we rely on
the '%package-name' variable for generating the version message, this
test is no longer reliable.

* gash/commands/ls.scm (*help-message*): New variable.
(*version-message*): New variable.
(*options-grammar*): New variable.
(ls): Parse options using 'parse-options'.
* tests/core-utils.org (ls-sanity): Remove test.