regex: Process special backslash escapes.
This used to happen in both AWK and Sed, but was lost in AWK when it
was converted to using the 'regex' module.  This change restores it in
AWK by adding it to the 'regex' module, and fixes Sed to not process
escapes a second time.

* gash-utils/regex.scm (*special-escapes*): New variable.
(special-escape?): New procedure.
(read-bracket-expression): Process special escapes.
(read-re-until): Process special escapes, with escaping the delimiter
taking precedence over them.
* gash/commands/sed.scm (replace-escapes): Remove procedure.
(substitute, address->pred): Do not use it.
* tests/unit/sed-reader.scm (Allows escaping address delimiter):
Adjust test.